ART-68 First Stage of Modularization achieved
